Prime Number Checker in Python
This Python program determines whether a given number is prime or not. By accepting user input, it checks if the number is divisible by any integers other than 1 and itself, using an optimized algorithm that iterates up to the square root of the number. The program returns "Prime" for prime numbers and "Not Prime" otherwise.